fix(schemas): set domain status enum guard
