+++ packages/schemas/src/foundations/jsonb-types.ts
@@ -269,3 +269,12 @@ export const cloudflareDataGuard = z
   .catchall(z.unknown());
 
 export type CloudflareData = z.infer<typeof cloudflareDataGuard>;
+
+export enum DomainStatus {
+  PendingVerification = 'PendingVerification',
+  PendingSsl = 'PendingSsl',
+  Active = 'Active',
+  Error = 'Error',
+}
+
+export const domainStatusGuard = z.nativeEnum(DomainStatus);
